- Confirm loanrequest added - calculator updated with templates - ShadowPaymentCommand added - LoanCommand added - LoanRequestCommand updated

+++ grails-app/controllers/commands/UserCommand.groovy
@@ -5,7 +5,7 @@ import enums.UserStatus
 import grails.validation.Validateable
 
 class UserCommand implements Validateable {
-    int id
+    Long id
     String firstName
     String lastName
     String username
